+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+Version 3.0.0-alpha3 *(2014-10-31)*
+-----------------------------------
+
+ * Use new icon names
+ * Implemented ResourceMenuItem properties and added Searchable property.
+
+Version 3.0.0-alpha2 *(2014-10-30)*
+-----------------------------------
+
+ * Added a FormHeader
+ * Allow all MenuItems to be passed in Resource::addItem, except for ResourceMenuItem
+
 Version 3.0.0-alpha1 *(2014-10-20)*
 -----------------------------------